+++ dynamic_color/example/lib/main.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 import 'common.dart';
 import 'complete_example.dart';
-import 'control_accent_color.dart';
+import 'accent_color.dart';
 import 'core_palette_visualization.dart';
 import 'dynamic_color_builder_example.dart';
 import 'get_core_palette_example.dart';
@@ -45,8 +45,8 @@ class ExampleApp extends StatelessWidget {
                   widget: CorePaletteVisualization(),
                 ),
                 const _ExampleAppButton(
-                  title: ControlAccentColorExample.title,
-                  widget: ControlAccentColorExample(),
+                  title: AccentColorExample.title,
+                  widget: AccentColorExample(),
                 ),
                 const _ExampleAppButton(
                   title: HarmonizationExample.title,